本番環境の準備をほぼ完了
ただしCredentials AuthorizationはDB接続で失敗する、OAuthに切り替える
