+++ docker-compose-prod.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-docker compose -f docker-compose-prod.yml up #--build
+docker compose -f docker-compose-prod.yml up --build
 
-docker cp sekirei-todo-nextjs:/app/.next ./next-prod
+#docker cp sekirei-todo-nextjs:/app/.next ./next-prod
 
-docker compose -f docker-compose-prod.yml down
+#docker compose -f docker-compose-prod.yml down
