+++ src/auth.ts
@@ -2,24 +2,10 @@ import NextAuth, { User } from "next-auth";
 import Credentials from 'next-auth/providers/credentials';
 import { authConfig } from './auth.config';
 import { z } from 'zod';
-import { db } from "@/db";
-import { eq } from 'drizzle-orm';
-import { users } from "@/db/schema";
 import { hashWithSalt } from '@/lib/crypto';
 
+import { getUser } from '@/actions/userActions';
 
-export async function getUser(username: string): Promise<{
-  username: string;
-  passWithSalt: string;
-} |undefined> {
-  try {
-    const user = await db.select().from(users).where(eq(users.id, username));
-    return { username: user[0].id, passWithSalt: user[0].passWithSalt };
-  } catch (error) {
-    console.error('Failed to fetch user:', error);
-    throw new Error(`Failed to fetch user: ${error}`);
-  }
-}
 
 export const { handlers: { GET, POST}, signIn, signOut, auth } = NextAuth({
   ...authConfig,
@@ -31,6 +17,9 @@ export const { handlers: { GET, POST}, signIn, signOut, auth } = NextAuth({
         password: {},
       },
       async authorize(credentials): Promise<User|null> {
+        'use server'
+
+        console.log('authorize()');
         const parsedCredentials = z
           .object({ username: z.string(), password: z.string() })
           .safeParse(credentials);
